x/mybank/types/message_my_mint_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/taiki-furu/mybank/testutil/sample"
)

func TestMsgMyMint_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgMyMint
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgMyMint{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgMyMint{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
